% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_function.R
\name{ca_function}
\alias{ca_function}
\title{A Cat Function}
\usage{
ca_function(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
