% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSampleDiffAnalysis.R
\name{calJSD}
\alias{calJSD}
\title{Title A function to calculate the JSD score.}
\usage{
calJSD(norm_data, save_path = NA, file_prefix)
}
\arguments{
\item{norm_data}{A normalized data obtained by **quantification**.}

\item{save_path}{The path to save the JSD score result.}

\item{file_prefix}{The save file prefix.}
}
\description{
Title A function to calculate the JSD score.
}
\examples{
 calJSDscore(norm_data = quant_data)

}
