\name{crimes}
\alias{crimes}
\title{U.S. National Crime Survey}
\description{Victimization status of households on two occasions.
}
\usage{data(crimes)}
\format{
  The matrix \code{crimes} contains the following columns:
  \describe{
    \item{V1}{Victimization status on first occasion (1=No, 2=Yes)}
    \item{V1}{Victimization status on second occasion (1=No, 2=Yes)}
    \item{Freq}{Count}  }
}

\source{
 Schafer (1996)  \emph{Analysis of Incomplete Multivariate Data.}
 Chapman & Hall, Section 7.4.3, which cites

 Kadane, J.B. (1985) Is victimization chronic? A Bayesian Analysis of
 multinomial missing data, \emph{Journal of Econometrics}, vol. 29, p. 47-67.
}
\keyword{datasets}
