% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{custom_order}
\alias{custom_order}
\title{Title  Order the character.}
\usage{
custom_order(x)
}
\arguments{
\item{x}{A vertor contains characters and numbers.}
}
\description{
Title  Order the character.
}
\examples{
 custom_order(x)
}
