% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoAccessibilityUtils.R
\name{estimate_distance_parameter}
\alias{estimate_distance_parameter}
\title{Title}
\usage{
estimate_distance_parameter(
  indata,
  window = window,
  maxit = 100,
  s = 0.75,
  sample_num = 100,
  distance_constraint = 250000,
  distance_parameter_convergence = 1e-22,
  max_elements = 200,
  genomic_coords = NULL
)
}
\arguments{
\item{genomic_coords}{}
}
\description{
Title
}
