% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoAccessibilityUtils.R
\name{find_distance_parameter}
\alias{find_distance_parameter}
\title{Title}
\usage{
find_distance_parameter(
  dist_mat,
  gene_range,
  maxit,
  null_rho,
  s,
  distance_constraint,
  distance_parameter_convergence
)
}
\arguments{
\item{distance_parameter_convergence}{}
}
\description{
Title
}
