% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoAccessibilityUtils.R
\name{generate_cicero_models}
\alias{generate_cicero_models}
\title{Title}
\usage{
generate_cicero_models(
  indata,
  distance_parameter,
  s = 0.75,
  window = 5e+05,
  max_elements = 200,
  genomic_coords = NULL
)
}
\arguments{
\item{genomic_coords}{}
}
\description{
Title
}
