% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialAnalysis.R
\name{getDiffTargetGenes}
\alias{getDiffTargetGenes}
\title{Title A function to call the differential peak target genes.}
\usage{
getDiffTargetGenes(diff_data, save_path = NA, file_prefix)
}
\arguments{
\item{diff_data}{The differential data obtained by **getDiffPeak**.}

\item{save_path}{The path to save the result.}

\item{file_prefix}{The save file prefix.}
}
\description{
Title A function to call the differential peak target genes.
}
\examples{
  getDiffTargetGenes(diff_data = diff_peak)

}
