% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifAnalysis.R
\name{getJasparMotifs2018}
\alias{getJasparMotifs2018}
\title{Title  getJasparMotifs2018}
\usage{
getJasparMotifs2018(species = "Homo sapiens", collection = "CORE", ...)
}
\arguments{
\item{species}{Spices.}

\item{collection}{Collection type.}

\item{...}{}
}
\description{
Title  getJasparMotifs2018
}
\examples{
 getJasparMotifs2018(species = "Homo sapiens", collection = "CORE")
}
