% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrice.R
\name{getPrice}
\alias{getPrice}
\title{Get daily price information}
\usage{
getPrice('035420')
}
\arguments{
\item{code}{stockCode}
}
\value{
vector with price, volume, low, high, changed, changed ratio
}
\description{
description
}
