% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperEnhancer.R
\name{getSuperEnhancer}
\alias{getSuperEnhancer}
\title{Title}
\usage{
getSuperEnhancer(
  sample_list,
  peak_path,
  peak_suffix = ".narrowPeak.bed",
  N_top = 1000,
  blacklist = NA
)
}
\arguments{
\item{sample_list}{A vectot contains sample names.}

\item{peak_path}{The path contains the sample peak files.}

\item{peak_suffix}{The suffix of peak files.}

\item{N_top}{The top number of peakset return.}

\item{blacklist}{A blacklist would prevent us from identifying super enhancers in those areas.}
}
\description{
Title
}
\examples{
 getSuperEnhancer(sample_list=c("s1","s2"),
                            peak_path="~/peak/",
                            blacklist="~/peak/balcklist.bed")

}
