% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{getTimeATAC}
\alias{getTimeATAC}
\title{Title}
\usage{
getTimeATAC(norm_data, Palette = NA, return_matrix = FALSE)
}
\arguments{
\item{norm_data}{Input the normalized dynamic matrix.}

\item{Palette}{Providing the custom Palette from the https://r-charts.com/color-palettes/.}
}
\description{
Title
}
