% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSampleDiffAnalysis.R
\name{getTopSpecifcPeaks}
\alias{getTopSpecifcPeaks}
\title{Title   getTopSpecifcPeaks}
\usage{
getTopSpecifcPeaks(
  spm_data,
  norm_data,
  top_N = 1000,
  save_path = NA,
  file_prefix = NA
)
}
\arguments{
\item{spm_data}{The Specificity Measure(SPM) result obtained by **calSPM**.}

\item{norm_data}{A normalized data obtained by **quantification**.}

\item{top_N}{The top N specific peaks to return.}

\item{save_path}{The path to save file.}

\item{file_prefix}{The save file prefix.}
}
\description{
Title   getTopSpecifcPeaks
}
\examples{
   getTopSpecifcPeaks(spm, quant_data, top_N = 1000)
}
