% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialAnalysis.R
\name{getTraidsTargetGenes}
\alias{getTraidsTargetGenes}
\title{Title A function to get the target genes of traids peak.}
\usage{
getTraidsTargetGenes(traids_data, save_path = NA, file_prefix = NA)
}
\arguments{
\item{traids_data}{The traids result obtained by **getTraids**.}

\item{save_path}{The path to save the result.}

\item{file_prefix}{The prefix of the file.}
}
\description{
Title A function to get the target genes of traids peak.
}
\examples{
  getTraidsTargetGenes(traids_data = res)
}
