% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASEnrichment.R
\name{GWASNorm}
\alias{GWASNorm}
\title{Title}
\usage{
GWASNorm(indir, outdir, prefix, drop_quantile = 0.1)
}
\arguments{
\item{indir}{The directory containing the counts files obtained by the **getCountSplit** function.}

\item{outdir}{The directory save the normalized data.}

\item{prefix}{The prefix of save files.}
}
\description{
Title
}
\examples{
   GWASNorm(indir = "./gwas/indir", outdir = "./gwas_out", prefix = "Immune_Cell")
}
