% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitten.R
\name{kitten}
\alias{kitten}
\title{Download kittens}
\usage{
kitten(width, height, destfile, ...)
}
\arguments{
\item{width}{Width of the kitten, in pixels}

\item{height}{Height of the kitten, in pixels}

\item{destfile}{Place to download the kitten to.  \code{\link{tempfile}()} makes a good place}

\item{...}{Additional arguments passed through to \code{\link{download.file}}}
}
\value{
Returns the filename of the kitten (i.e., \code{destfile})
}
\description{
Download kittens from the internet
}
\details{
This is the details section
}
\examples{
# download a small cat
file <- kitten(100, 100, tempfile())
# download a very large cat
file <- kitten(1000, 100, tempfile())
}

