% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mailMe.R
\name{mailMe}
\alias{mailMe}
\title{Generate / Send daily Report}
\usage{
mailMe()
}
\value{
NO RETURN
}
\description{
creds_key with id = 'gmail' should be declared before
}
\seealso{
blastula::create_smtp_creds_key
}
