% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifAnalysis.R
\name{makeSummarizedExperiment}
\alias{makeSummarizedExperiment}
\title{Title  Make SummarizedExperiment for chromVAR input.}
\usage{
makeSummarizedExperiment(count_data)
}
\arguments{
\item{count_data}{The raw count data obtained by **getCounts**.}
}
\description{
Title  Make SummarizedExperiment for chromVAR input.
}
\examples{
makeSummarizedExperiment(count_data)

}
