% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASEnrichment.R
\name{plotGER}
\alias{plotGER}
\title{Title}
\usage{
plotGER(result, plot_type = "lollipop")
}
\arguments{
\item{result}{The GWAS enrichment result obtained by **getGER**.}

\item{plot_type}{c("radar","lollipop").}
}
\description{
Title
}
