% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSampleDiffAnalysis.R
\name{plotGONetwork}
\alias{plotGONetwork}
\title{Title   A function to compare the GO enrichment result with Network.}
\usage{
plotGONetwork(
  target,
  orgdb,
  ont = "BP",
  showCategory = 50,
  category_node = 0.5,
  line = 0.5,
  category_label = 1.2
)
}
\arguments{
\item{target}{The target genes result obtained by **getClusterTargetGenes**.}

\item{orgdb}{The orgdb annotation package name.}

\item{ont}{The GO type to enrich. Three type are available: "BP", "CC", "MP". Default is BP.}

\item{showCategory}{The max category number to show. Default is 50.}

\item{category_node}{The size of category nodes. Default is 0.5.}

\item{line}{The width of the line. Default is 0.5.}

\item{category_labelThe}{size of category label. Default is 1.2.}
}
\description{
Title   A function to compare the GO enrichment result with Network.
}
\examples{
  plotGONetwork(target = res, orgdb = org.Hs.eg.db)

}
