% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialAnalysis.R
\name{plotTraids}
\alias{plotTraids}
\title{Title   A function to plot the traids result.}
\usage{
plotTraids(traids_data)
}
\arguments{
\item{traids_data}{The traids result obtained by **getTraids**.}
}
\description{
Title   A function to plot the traids result.
}
\examples{
 plotTraids(traids_data = res)
}
