% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVenn.R
\name{plotVenn}
\alias{plotVenn}
\title{Draw venn plot of bed intersection}
\usage{
plotVenn(
  input_df,
  bed_list,
  colors = c("#4E79A7FF", "#F28E2BFF", "#E15759FF", "#76B7B2FF", "#59A14FFF",
    "#EDC948FF", "#B07AA1FF"),
  opacity = 0.8,
  plotsize = 15
)
}
\arguments{
\item{input_df}{A dataframe contains key and value(From get_combain_result())}

\item{bed_list}{A list contains records of all bed files(Read by function get_beds())}

\item{colors}{Default is Tableau_10}

\item{opacity}{Default is 0.3}

\item{plotsize}{Default is 15}
}
\value{
Ggplot object of venn
}
\description{
This function facilitates the generation of Venn plot based on the intersection
results obtained from the get_beds() function and get_combine_result()
}
\examples{
# Get beds using get_beds
beds <- get_beds(c('path1', 'path2', ... ))

# Use get_combine_result() to get intersection result
# The processing time may vary depending on the number of sets involved
inputs <- get_combine_result(beds)

draw_venn(inputs, beds)
}
