\name{RiquelmeTaxonomy}
\alias{RiquelmeTaxonomy}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Taxonomy table for the Riquelme dataset
}

\description{
  Taxonomy table for the Riquelme dataset.
}

\source{
Riquelme E, Zhang Y, Zhang L, Montiel M, Zoltan M, Dong W, Quesada P, Sahin I, Chandra V, San Lucas A, Scheet P, Xu H, Hanash SM, Feng L, Burks JK, Do KA, Peterson CB, Nejman D, Tzeng CD, Kim MP, Sears CL, Ajami N, Petrosino J, Wood LD, Maitra A, Straussman R, Katz M, White JR, Jenq R, Wargo J, McAllister F. Tumor Microbiome Diversity and Composition Influence Pancreatic Cancer Outcomes. Cell. 2019 Aug 8;178(4):795-806.e12. doi: 10.1016/j.cell.2019.07.008. PMID: 31398337; PMCID: PMC7288240.
}
