% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoAccessibilityUtils.R
\name{run_cicero}
\alias{run_cicero}
\title{Title}
\usage{
run_cicero(
  indata,
  genomic_coords,
  window = 45000,
  silent = FALSE,
  sample_num = 100
)
}
\arguments{
\item{sample_num}{}
}
\description{
Title
}
