\name{slice.index}
\alias{slice.index}
\title{Slice.index}
\description{Mimics the behaviour of the S-Plus function \code{slice.index}}
\note{Code originally from Peter Dalgaard, answering a query from
 Fernando Tusell on the R list (30.March.2000). Recent versions
 of R have this function built-in.
}
\keyword{misc}
% Converted by Sd2Rd version 1.21.

