% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFrgn.R
\name{getFrgn}
\alias{getFrgn}
\title{Get Foreign / Organization trend information}
\usage{
getFrgn('035420')
}
\arguments{
\item{code}{stockCode}

\item{day}{how much day max 20}
}
\value{
data.frame with
}
\description{
this will not include last day's information.
}
