% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintAnalysis.R
\name{plotDiffFootprint}
\alias{plotDiffFootprint}
\title{Title}
\usage{
plotDiffFootprint(
  samples,
  cutoff = NA,
  plot_type = c("Volcano", "Bar", "Lollipop"),
  color = NA,
  save_path = NA,
  file_prefix = NA,
  figure_width = 8.21,
  figure_height = 8.21
)
}
\arguments{
\item{figure_height}{}
}
\description{
Title
}
\examples{
  plotDiffFootprint(samples=c("NIP_YP1","NIP_YP2"), bindetect_path="/public/workspace/zhutao/encode/CAT",plot_type="Bar",cutoff=0.1)

}
